+++ R/sim_adrf.R
@@ -88,10 +88,10 @@ sim_adrf <- function(sim,
   contrast <- match_arg(contrast, c("adrf", "amef"))
 
   if (is_misim) {
-    dat <- do.call("rbind", lapply(sim$fit, insight::get_predictors))
+    dat <- do.call("rbind", lapply(sim$fit, insight::get_predictors, verbose = FALSE))
   }
   else {
-    dat <- insight::get_predictors(sim$fit)
+    dat <- insight::get_predictors(sim$fit, verbose = FALSE)
   }
 
   if (!var %in% names(dat)) {